only delete resources if resources are left over

+++ .github/actions/e2e_cleanup_timeframe/e2e-cleanup.sh
@@ -13,10 +13,12 @@ function download_tfstate_artifact {
 
 # delete_resources runs terraform destroy on the constellation-terraform subfolder of a given folder.
 function delete_resources {
-  cd "$1/constellation-terraform" || exit 1
-  terraform init > /dev/null || exit 1 # first, install plugins
-  terraform destroy -auto-approve || exit 1
-  cd ../../ || exit 1
+  if [ -d "$1/constellation-terraform" ]
+    cd "$1/constellation-terraform" || exit 1
+    terraform init > /dev/null || exit 1 # first, install plugins
+    terraform destroy -auto-approve || exit 1
+    cd ../../ || exit 1
+  fi
 }
 
 # delete_iam_config runs terraform destroy on the constellation-iam-terraform subfolder of a given folder.
